chore: move sqlachemy to being a dev only dependency
