+++ requirements.dev.in
@@ -7,3 +7,4 @@ pre-commit
 pytest
 pytest-freezegun
 pytest-subtests
+sqlalchemy
