+++ requirements.dev.txt
@@ -87,6 +87,8 @@ six==1.14.0
     #   -c requirements.prod.txt
     #   python-dateutil
     #   virtualenv
+sqlalchemy==1.3.15
+    # via -r requirements.dev.in
 toml==0.10.2
     # via
     #   pre-commit
