+++ requirements.prod.txt
@@ -65,8 +65,6 @@ six==1.14.0
     # via
     #   presto-python-client
     #   python-dateutil
-sqlalchemy==1.3.15
-    # via opensafely-cohort-extractor (setup.py)
 sqlparse==0.3.1
     # via opensafely-cohort-extractor (setup.py)
 structlog==20.2.0
